transcendence/backend/requirements.txt
Django==5.0.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
django-prometheus==2.3.1
django-redis>=5.4.0
redis>=5.0.0
daphne==4.1.0
channels==4.0.0
channels-redis==4.2.0
python-chess==1.999
requests>=2.31.0